stage5 new design

+++ lib/features/stage5_1_missing_weight/domain/stage5_price_form.dart
@@ -9,8 +9,8 @@ abstract class Stage5PriceFormData with _$Stage5PriceFormData {
     required String id,
     required String projectId,
     required DateTime date,
-    @Default(0) int pricePerKilo,
-    @Default(0) int installment,
+    @Default(0) double pricePerKilo,
+    @Default(0) double installment,
   }) = _Stage5PriceFormData;
 
   factory Stage5PriceFormData.fromJson(Map<String, dynamic> json) =>
